Combined both apps into one
