+++ package.json
@@ -9,6 +9,7 @@
     "lucide-react": "^0.439.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
