+++ src/MacronutrientAnalyzer.css
@@ -1,0 +1,109 @@
+.analyzer-container {
+    display: flex;
+    max-width: 1200px;
+    margin: 0 auto;
+  }
+  
+  .left-panel {
+    flex: 1;
+    background-color: #fff;
+    padding: 20px;
+    border-radius: 5px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    margin-right: 20px;
+  }
+  
+  .right-panel {
+    position: fixed; /* Fixed position */
+    right: 20px;
+    top: 20px;
+    width: 300px; /* More compact width */
+    background-color: #fff;
+    padding: 15px;
+    border-radius: 5px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    max-height: 90vh; /* Maintain space on the page */
+    overflow-y: auto; /* Scroll if content is too long */
+  }
+  
+  h1 {
+    text-align: center;
+    color: #333;
+  }
+  
+  #searchInput {
+    width: 100%;
+    padding: 10px;
+    margin-bottom: 20px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+  }
+  
+  #results {
+    margin-top: 20px;
+  }
+  
+  .food-item {
+    background-color: #f9f9f9;
+    padding: 10px;
+    margin-bottom: 10px;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+  }
+  
+  .food-item:hover {
+    background-color: #e9e9e9;
+  }
+  
+  .food-item h3 {
+    margin-top: 0;
+    color: #333;
+  }
+  
+  .macro {
+    display: block;
+    margin-bottom: 3px;
+    font-size: 0.9em; /* Smaller font for compactness */
+  }
+  
+  .score {
+    font-size: 1em;
+    font-weight: bold;
+    padding: 5px 8px;
+    border-radius: 4px;
+    display: inline-block;
+  }
+  
+  .score-good {
+    background-color: #4caf50;
+    color: white;
+  }
+  
+  .score-bad {
+    background-color: #f44336;
+    color: white;
+  }
+  
+  .score-medium {
+    background-color: #ffc107;
+    color: white;
+  }
+  
+  #showMoreBtn {
+    display: block;
+    width: 100%;
+    padding: 10px;
+    background-color: #4caf50;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 16px;
+    margin-top: 10px;
+  }
+  
+  #showMoreBtn:hover {
+    background-color: #45a049;
+  }
+
